Export mock context class as well as create context functions
