+++ utils/createMockActionContext.js
@@ -5,40 +5,16 @@
 'use strict';
 
 var dispatchr = require('dispatchr');
-
-function MockActionContext (dispatcherContext) {
-    this.dispatcherContext = dispatcherContext;
-    this.executeActionCalls = [];
-    this.dispatchCalls = [];
-}
-
-MockActionContext.prototype.getStore = function (name) {
-    return this.dispatcherContext.getStore(name);
-};
-
-MockActionContext.prototype.dispatch = function (name, payload) {
-    this.dispatchCalls.push({
-        name: name,
-        payload: payload
-    });
-    this.dispatcherContext.dispatch(name, payload);
-};
-
-MockActionContext.prototype.executeAction = function (action, payload, callback) {
-    this.executeActionCalls.push({
-        action: action,
-        payload: payload
-    });
-    action(this, payload, callback);
-};
+var MockActionContextClass = require('./MockActionContext');
 
 module.exports = function createMockActionContext(options) {
     options = options || {};
+    options.mockActionContextClass = options.mockActionContextClass || MockActionContextClass;
     options.stores = options.stores || [];
     options.dispatcher = options.dispatcher || dispatchr.createDispatcher({
         stores: options.stores
     });
     options.dispatcherContext = options.dispatcherContext || options.dispatcher.createContext();
 
-    return new MockActionContext(options.dispatcherContext);
+    return new options.mockActionContextClass(options.dispatcherContext);
 };
